object detection/torchserve/requirements_torchserve.txt
torch
torchvision
opencv-python
numpy
Pillow
